+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 -   Proper error handling for invalid external module code.
 -   **call-fns** command for calling a Rhai function. Note: The function can only see poll/listen variables as their initial value.
 -   **update** command with **--inject-vars** flag to update widget state. Note: All poll/listen variables will reset to their initial values.
+-   `std::command` module for running shell commands.
 
 ### Changed